Add files via upload
This is a Data Analysis project on paying overtime to employees. In the dataset, overtime hour is calculated using the 'IF' formula taking 40 h/ week as the base hour. The overtime bonus is calculated and the total payment to be made to the employee is done. The maximum, minimum, average, and total of overtime hours and total payment is done.